+++ main.go
@@ -21,7 +21,7 @@ import (
 
 const (
 	// appName    = "KrankyBear Reader"
-	appVersion = "0.1.0" // see FyneApp.toml
+	appVersion = "0.2.0" // see FyneApp.toml
 	appAuthor  = "Allan Marillier"
 	appID      = "com.github.amarillier.KrankyBearReader"
 )